Console application with a method named PrimeFactors
Prime factors are the combination of the smallest prime numbers, that, when multiplied together, will produce the original number. Consider the following example:

Prime factors of 4 are: 2 x 2

Prime factors of 7 are: 7

Prime factors of 30 are: 2 x 3 x 5

Prime factors of 40 are: 2 x 2 x 2 x 5

Prime factors of 50 are: 2 x 5 x 5

Create a console application with a method named PrimeFactors that, when passed an int variable as a parameter, returns a string showing its prime factors.

Use the debugging tools and write unit tests to ensure that your function works correctly with multiple inputs and returns the correct output.